refactor as a monorepo, using npm workspaces
