+++ .github/workflows/_common_jobs.yml
@@ -1,0 +1,35 @@
+name: Launch CI jobs for a package or app
+on:
+  workflow_call:
+    inputs:
+      workspace:
+        required: true
+        type: string
+jobs:
+  # lint:
+  #   runs-on: ubuntu-latest
+  #   steps:
+  #     - uses: actions/checkout@v4
+  #     - run: npm i
+  #     - run: npm run lint -w ${{ inputs.workspace }}
+
+  typecheck:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v4
+      - run: npm i
+      - run: npm run typecheck -w ${{ inputs.workspace }}
+
+  # test:
+  #   runs-on: ubuntu-latest
+  #   steps:
+  #     - uses: actions/checkout@v4
+  #     - run: npm i
+  #     - run: npm run coverage -w ${{ inputs.workspace }}
+
+  buildcheck:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v4
+      - run: npm i
+      - run: npm run build -w ${{ inputs.workspace }}
