+++ .github/workflows/deploy.yml
@@ -7,11 +7,8 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v4
-      - uses: actions/setup-node@v4
-        with:
-          node-version: "22.x"
-      - run: npm ci
-      - run: npm run build:app
+      - run: npm i
+      - run: npm run build -w @source-cooperative/components-demo
       - name: Upload static files as artifact
         id: deployment
         uses: actions/upload-pages-artifact@v3
