+++ package.json
@@ -1,58 +1,27 @@
 {
-	"name": "@source-cooperative/components",
-	"version": "1.0.15",
-	"description": "Source Cooperative React Components",
-	"main": "dst/index.js",
-	"module": "dst/index.esm.js",
-	"repository": {
-		"type": "git",
-		"url": "git+https://github.com/source-cooperative/components.git"
-	},
-	"keywords": [
-		"theme-ui",
-		"source-cooperative"
-	],
-	"files": [
-		"dst"
-	],
-	"author": "radiantearth",
-	"license": "MIT",
-	"bugs": {
-		"url": "https://github.com/source-cooperative/components/issues"
-	},
-	"dependencies": {
-		"@theme-ui/color": "0.16.1",
-		"@theme-ui/prism": "0.16.1",
-		"react-hot-toast": "^2.4.1",
-		"react-loading-skeleton": "^3.3.1",
-		"usehooks-ts": "^2.9.1",
-		"@babel/plugin-proposal-decorators": "7.23.3",
-		"babel-preset-gatsby": "3.12.1"
-	},
-	"peerDependencies": {
-		"@types/node": "20.9.4",
-		"@types/react": "18.2.38",
-		"next": "13.5.6",
-		"theme-ui": "0.16.1",
-		"@emotion/react": "11.11.1",
-		"react": ">=18",
-		"react-dom": ">=18"
-	},
-	"devDependencies": {
-		"@source-cooperative/theme": "1.0.11",
-		"@types/node": "20.9.4",
-		"@types/react": "18.2.38",
-		"prettier": "2.3.0",
-		"typescript": "^5.2.2",
-		"microbundle": "^0.13.0",
-		"rimraf": "3.0.2"
-	},
-	"homepage": "https://github.com/source-cooperative/components#readme",
-	"scripts": {
-		"dev": "next dev",
-		"format": "prettier --write 'index.js'",
-		"build:app": "next build",
-		"build:package": "rimraf dst && microbundle src/index.tsx -o dst/index.js --jsx React.createElement -f modern,es,cjs --jsxFragment React.Fragment",
-		"prepare": "npm run build:package"
-	}
+  "name": "@source-cooperative/monorepo",
+  "version": "0.0.0",
+  "private": true,
+  "license": "MIT",
+  "workspaces": [
+    "packages/components",
+    "apps/components-demo"
+  ],
+  "scripts": {
+    "build": "npm run build -ws"
+  },
+  "overrides": {
+    "@emotion/react": "11.11.1",
+    "next": "13.5.6",
+    "react": "18.3.1",
+    "react-dom": "18.3.1",
+    "react-loading-skeleton": "^3.3.1",
+    "theme-ui": "0.16.1"
+  },
+  "devDependencies": {
+    "@types/node": "20.9.4",
+    "@types/react": "18.2.38",
+    "prettier": "2.3.0",
+    "typescript": "^5.2.2"
+  }
 }
