+++ packages/components/src/components/Footer.tsx
