+++ packages/components/src/components/SearchBar.tsx
